+++ PluginInstallation/CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased](https://github.com/packlink-dev/prestashop_module/compare/master...logeecom:dev)
+### Added
+- Added compatibility with PrestaShop version 1.7.7.
+
 ## [v3.0.4](https://github.com/packlink-dev/prestashop_module/compare/v3.0.2...v3.0.4) - 2020-11-10
 ### Changed
 - Update to the latest core version 3.0.6.